split disc hashing out of hash.c into hash_disc.c
