+++ test/rcheevos-test.vcxproj
@@ -159,6 +159,7 @@
     <ClCompile Include="..\src\rhash\aes.c" />
     <ClCompile Include="..\src\rhash\cdreader.c" />
     <ClCompile Include="..\src\rhash\hash.c" />
+    <ClCompile Include="..\src\rhash\hash_disc.c" />
     <ClCompile Include="..\src\rhash\hash_encrypted.c" />
     <ClCompile Include="..\src\rhash\hash_rom.c" />
     <ClCompile Include="..\src\rhash\hash_zip.c" />
@@ -187,6 +188,7 @@
     <ClCompile Include="rhash\mock_filereader.c" />
     <ClCompile Include="rhash\test_cdreader.c" />
     <ClCompile Include="rhash\test_hash.c" />
+    <ClCompile Include="rhash\test_hash_disc.c" />
     <ClCompile Include="rhash\test_hash_rom.c" />
     <ClCompile Include="rhash\test_hash_zip.c" />
     <ClCompile Include="rurl\test_url.c" />
