+++ test/rcheevos-test.vcxproj.filters
@@ -224,6 +224,12 @@
     <ClCompile Include="rhash\test_hash_rom.c">
       <Filter>tests\rhash</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\rhash\hash_disc.c">
+      <Filter>src\rhash</Filter>
+    </ClCompile>
+    <ClCompile Include="rhash\test_hash_disc.c">
+      <Filter>tests\rhash</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\rcheevos.h">
