+++ test/rhash/mock_filereader.h
@@ -10,7 +10,11 @@ RC_BEGIN_C_DECLS
 
 void init_mock_filereader();
 void get_mock_filereader(struct rc_hash_filereader* reader);
+
+#ifndef RC_HASH_NO_DISC
 void init_mock_cdreader();
+void mock_cd_num_tracks(int num_tracks);
+#endif
 
 void rc_hash_reset_filereader();
 
@@ -19,8 +23,6 @@ void mock_file_text(int index, const char* filename, const char* contents);
 void mock_empty_file(int index, const char* filename, size_t mock_size);
 void mock_file_size(int index, size_t mock_size);
 void mock_file_first_sector(int index, int first_sector);
-
-void mock_cd_num_tracks(int num_tracks);
 
 const char* get_mock_filename(void* file_handle);
